SMP: introduced SMP initialization and booting for additional CPUs. Dev: added device management structures and functions for device registration and IRQ handling. Dev: added device IDs as a uint64 in the format major.device.minor Timer: Implemented timer registration, and management. Kstack: Created stack allocator for kernel stacks with guard pages. Kerrno: Added new kerrno based on POSIX errno. Stack_info: Added for POSIX later. Sync: Added spinlocks for post 1.0 kernel.

+++ src/kernel/arch/x86/include/kernel/acpi/apic/ioapic.h
@@ -23,7 +23,7 @@ typedef struct iso_entry
     interrupt_vector_t vector;
     gsi_t gsi;
     uint16_t flags;
-    uint8_t lapic_target;
+    uint32_t lapic_target;
 } iso_entry_t;
 
 typedef struct ioapic_entry
@@ -55,4 +55,4 @@ uint8_t ioapic_get_id(gsi_t gsi);
 void ioapic_unmask(uint8_t ioapic_id, uint32_t irq);
 void ioapic_mask(uint8_t ioapic_id, uint32_t irq);
 uint32_t ioapic_is_mask(uint8_t ioapic_id, uint32_t irq);
-void ioapic_set_entry(uint8_t ioapic_id, gsi_t gsi, uint8_t vector, uint16_t flags, uint8_t dest_apic_id);
+void ioapic_set_entry(uint8_t ioapic_id, gsi_t gsi, uint8_t vector, uint16_t flags, uint32_t dest_apic_id);
